% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-methods.R
\name{autoplot.egfrtraj_single}
\alias{autoplot.egfrtraj_single}
\title{Autoplot method for egfrtraj_single objects}
\usage{
\method{autoplot}{egfrtraj_single}(object, ...)
}
\arguments{
\item{object}{An egfrtraj_single object}

\item{...}{Additional arguments passed to ggplot2::ggplot()}
}
\value{
A ggplot object
}
\description{
Returns a ggplot object of the eGFR trajectory with selected model fit,
prediction intervals, and breakpoint line (if applicable).
}
