% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{has_segmented}
\alias{has_segmented}
\title{Check if segmented package is available and loaded}
\usage{
has_segmented()
}
\value{
Logical
}
\description{
Check if segmented package is available and loaded
}
\keyword{internal}
