% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prefer_tidyverbs.R
\name{prefer_tidyverbs}
\alias{prefer_tidyverbs}
\title{Prefer common tidyverse verbs (dplyr + tidyr)}
\usage{
prefer_tidyverbs()
}
\description{
Gives priority to dplyr and tidyr verbs, including select/rename/distinct.
}
\examples{
\dontrun{
library(conflicted)
library(dplyr)
library(tidyr)
library(segmented)
library(egfrtraj)
data(example_egfr_data)
egfrtraj::prefer_tidyverbs()

# Now filter(), lag(), select() etc. will use dplyr versions
example_egfr_data |> filter(time > 0)
}
}
