% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print-methods.R
\name{summary.egfr_traj}
\alias{summary.egfr_traj}
\title{Summarise a single eGFR trajectory object}
\usage{
\method{summary}{egfr_traj}(object, ...)
}
\arguments{
\item{object}{An \code{egfr_traj} object}

\item{...}{Additional arguments (currently ignored)}
}
\value{
A one-row tibble with key summary statistics
}
\description{
Summarise a single eGFR trajectory object
}
