% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print-methods.R
\name{summary.egfr_traj_list}
\alias{summary.egfr_traj_list}
\title{Summarise results from classify_multiple_trajectories}
\usage{
\method{summary}{egfr_traj_list}(object, ...)
}
\arguments{
\item{object}{Named list of \code{egfr_traj} objects}

\item{...}{Additional arguments (currently ignored)}
}
\value{
A tibble with one row per subject
}
\description{
Summarise results from classify_multiple_trajectories
}
