% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{classify_multiple_trajectories}
\alias{classify_multiple_trajectories}
\title{Classify multiple eGFR trajectories}
\usage{
classify_multiple_trajectories(
  data,
  id_col = "id",
  time_col = "time",
  egfr_col = "egfr",
  add_cols = NULL,
  bic_tie_threshold = 4,
  ci_level = 0.95,
  best_model_only = TRUE,
  estimates_only = FALSE
)
}
\arguments{
\item{data}{Data frame with multiple (or single) subjects}

\item{id_col, time_col, egfr_col}{Column names}

\item{add_cols}{Character vector of additional columns to return separately}

\item{bic_tie_threshold}{Default 4}

\item{ci_level}{Default 0.95}

\item{best_model_only}{If \code{TRUE} (default), keep only selected model results}

\item{estimates_only}{If \code{TRUE}, drop standard error columns}
}
\value{
Named list of \code{egfr_traj} objects (one per subject)
}
\description{
Classify multiple eGFR trajectories
}
