% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-methods.R
\name{plot.egfr_traj}
\alias{plot.egfr_traj}
\title{Plot an eGFR trajectory}
\usage{
\method{plot}{egfr_traj}(x, fitted = "best", show_ci = TRUE, ...)
}
\arguments{
\item{x}{An \code{egfr_traj} object}

\item{fitted}{Character string: \code{"best"} (default), \code{"lin"}, \code{"quad"}, \code{"seg"}, or \code{NULL} (observed only).}

\item{show_ci}{Logical. Show CI ribbon for the fitted model? Default \code{TRUE}.}

\item{...}{Passed to ggplot2}
}
\value{
ggplot object
}
\description{
Plot an eGFR trajectory
}
